requirements.txt
slack-sdk==3.26.1
python-dotenv==1.0.0 